# Improve Array.Sort(T[]) performance