@damageboy 2020-04-24
Issue seems resolved.
Comapred 3.1.201, master without PR, PR branch.
perf now seems better than 3.1.201 in the higher sort problem sizes:

|    Method | Toolchain          |        N |    Mean [us] | Error [us] |StdDev [us] | Time / N [ns] |
|---------- |--------------------|--------- |-------------:|-----------:|-----------:|--------------:|
| ArraySort | `3.1.201`          |      100 |       0.9818 |     0.0152 |     0.0213 |     9.8177 |
| ArraySort | `master/08285b1`   |      100 |       1.1140 |     0.0230 |     0.0384 |    11.1398 |
| ArraySort | `pr/35297/2eef7dd` |      100 |       1.0903 |     0.0215 |     0.0287 |    10.9032 |
| ArraySort | `3.1.201`          |     1000 |      16.9130 |     1.0379 |     1.5213 |    16.9130 |
| ArraySort | `master/08285b1`   |     1000 |      21.9628 |     0.4283 |     0.6143 |    21.9628 |
| ArraySort | `pr/35297/2eef7dd` |     1000 |      17.2827 |     0.3448 |     0.5665 |    17.2827 |
| ArraySort | `3.1.201`          |    10000 |     440.7663 |     1.1282 |     1.5816 |    44.0766 |
| ArraySort | `master/08285b1`   |    10000 |     507.5048 |     1.4836 |     1.2388 |    50.7505 |
| ArraySort | `pr/35297/2eef7dd` |    10000 |     457.2838 |     9.0489 |     8.0216 |    45.7284 |
| ArraySort | `3.1.201`          |   100000 |   5,734.5160 |     5.6595 |     8.2957 |    57.3452 |
| ArraySort | `master/08285b1`   |   100000 |   6,474.4154 |    57.1762 |    53.4827 |    64.7442 |
| ArraySort | `pr/35297/2eef7dd` |   100000 |   5,648.4342 |    64.5477 |    60.3780 |    56.4843 |
| ArraySort | `3.1.201`          |  1000000 |  67,353.2352 |    52.7210 |    72.1651 |    67.3532 |
| ArraySort | `master/08285b1`   |  1000000 |  77,408.7551 |   580.1940 |   542.7138 |    77.4088 |
| ArraySort | `pr/35297/2eef7dd` |  1000000 |  65,691.5563 |   173.6212 |   162.4054 |    65.6916 |
| ArraySort | `3.1.201`          | 10000000 | 775,940.1604 |   693.8563 |   949.7583 |    77.5940 |
| ArraySort | `master/08285b1`   | 10000000 | 898,762.7196 | 2,351.3387 | 2,199.4437 |    89.8763 |
| ArraySort | `pr/35297/2eef7dd` | 10000000 | 749,980.9434 | 2,472.2225 | 2,064.4178 |    74.9981 |