@nietras 2020-04-27
@stephentoub @jkotas guess this means https://github.com/dotnet/corefx/pull/26859 could have been merged anyhow :)