@stephentoub 2020-04-27
> guess this means dotnet/corefx#26859 could have been merged anyhow :)

There's a ton more unsafe code in that PR than in this one.  The "unsafety" in this PR is scoped to just two functions, with extra scrutiny as called out in https://github.com/dotnet/corefx/pull/26859#issuecomment-451565437.